remove extra details from handlers using defaulted parameters
