+++ src/main/scala/com/gu/autoCancel/AutoCancelHandler.scala
@@ -1,19 +1,12 @@
 package com.gu.autoCancel
 
-import java.io.{InputStream, OutputStream}
-
-import com.amazonaws.services.lambda.runtime.Context
 import com.gu.effects.Logging
 import com.gu.util.apigateway.ApiGatewayHandler
 
 object AutoCancelHandler extends App with Logging {
 
-  def handleRequest(inputStream: InputStream, outputStream: OutputStream, context: Context): Unit = {
+  def handleRequest = ApiGatewayHandler() {
+    AutoCancelSteps.performZuoraAction
+  }_
 
-    val stage = System.getenv("Stage")
-
-    ApiGatewayHandler.handleRequest(inputStream, outputStream, context, stage) {
-      AutoCancelSteps.performZuoraAction
-    }
-  }
 }
