+++ src/main/scala/com/gu/effects/AwsS3.scala
@@ -3,10 +3,10 @@ package com.gu.effects
 import com.amazonaws.auth._
 import com.amazonaws.auth.profile.ProfileCredentialsProvider
 import com.amazonaws.services.s3.AmazonS3Client
-import com.amazonaws.services.s3.model.{GetObjectRequest, S3ObjectInputStream}
+import com.amazonaws.services.s3.model.{ GetObjectRequest, S3ObjectInputStream }
 
 import scala.io.Source
-import scala.util.{Failure, Try}
+import scala.util.{ Failure, Try }
 
 object ConfigLoad extends Logging {
